+++ semo-api/src/main/java/sandbox/semo/application/company/service/CompanyServiceImpl.java
@@ -1,14 +1,13 @@
 package sandbox.semo.application.company.service;
 
+import static sandbox.semo.application.company.exception.CompanyErrorCode.FORM_DOES_NOT_EXIST;
 import static sandbox.semo.application.company.exception.CompanyErrorCode.STATUS_NOT_APPROVED;
-import static sandbox.semo.application.form.exception.CompanyFormErrorCode.FORM_DOES_NOT_EXIST;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sandbox.semo.application.company.exception.CompanyBusinessException;
-import sandbox.semo.application.form.exception.CompanyFormBusinessException;
 import sandbox.semo.domain.common.entity.FormStatus;
 import sandbox.semo.domain.company.entity.Company;
 import sandbox.semo.domain.company.entity.CompanyForm;
@@ -27,7 +26,7 @@ public class CompanyServiceImpl implements CompanyService {
     @Transactional
     public Long companyRegister(Long formId) {
         CompanyForm companyForm = companyFormRepository.findById(formId)
-                .orElseThrow(() -> new CompanyFormBusinessException(FORM_DOES_NOT_EXIST));
+                .orElseThrow(() -> new CompanyBusinessException(FORM_DOES_NOT_EXIST));
 
         if (!companyForm.getFormStatus().equals(FormStatus.APPROVED)) {
             throw new CompanyBusinessException(STATUS_NOT_APPROVED);
